// frontend/src/generated/model/problemResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Polyglot V2 API
 * OpenAPI spec version: 0.1.0
 */
import type { ProblemResponseDetails } from "./problemResponseDetails";
import type { ProblemResponseFieldErrors } from "./problemResponseFieldErrors";

export interface ProblemResponse {
  code: string;
  correlation_id: string;
  detail: string;
  details?: ProblemResponseDetails;
  field_errors?: ProblemResponseFieldErrors;
  instance: string;
  message_key: string;
  request_id: string;
  retryable: boolean;
  status: number;
  title: string;
  type: string;
}